+++ setup.sh
@@ -1,15 +1,18 @@
 # Test of a fixed configure as bash shell
 echo $PWD
 
+export PEGASUS_PLATFORM=LINUX_X86_64_GNU
+
 # Set Root to the current directory.
 export ROOT=$PWD
 export PEGASUS_ROOT=$ROOT/pegasus
 export PEGASUS_HOME=$ROOT/home
 export PATH=$PEGASUS_HOME/bin:$PATH
-mkdir $PEGASUS_HOME
+mkdir $PEGASUS_HOME -p
 export LD_LIBRARY_PATH=$PEGASUS_HOME/lib64:
 
-export PEGASUS_PLATFORM=LINUX_X86_64_GNU
+
+
 
 # Storage mode of the CIM repository created on initial build
 # This may be modified using cimconfig.