+++ docs/changes.rst
@@ -20,6 +20,13 @@ Status: Development
 
 Bugs
 
+* Add code to set the system name into Indication subscription instances
+  when getInstance and EnumerateInstance called in IndicationServices.
+  This property was left empty before meaning that the value of the
+  paths was not equal to the value of the corresponding property
+  components because the properties had left the sysname key of the
+  Filter and Handler cim objects empty.  (See issue #82)
+
 
 Enhancements